TF_Timeseries _Sample_forecasting_for_Driver_count - sample Rnn time series forecast for a sample retail driver
TF_Timeseries _Sample_forecasting_for_Driver_count - Sample Rnn time series forecast for a sample retail driver
A retail driver that could be used to track invoices/reciepts or transaction is an ideal example for this . 

This driver can be forecasted based on day and based on the location. In the example shown, it is being predicted by day and by store for a single location. The code can be extended to multiple locations as well .